+++ lib/ts/recipe/emailpassword/index.ts
@@ -47,39 +47,6 @@ export default class Wrapper {
         return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.resetPasswordUsingToken({ token, newPassword });
     }
 
-    /**
-     * @deprecated Use supertokens.getUsersOldestFirst(...) function instead IF using core version >= 3.5
-     *   */
-    static getUsersOldestFirst(
-        limit?: number,
-        nextPaginationToken?: string
-    ): Promise<{
-        users: User[];
-        nextPaginationToken?: string | undefined;
-    }> {
-        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getUsersOldestFirst({ limit, nextPaginationToken });
-    }
-
-    /**
-     * @deprecated Use supertokens.getUsersNewestFirst(...) function instead IF using core version >= 3.5
-     *   */
-    static getUsersNewestFirst(
-        limit?: number,
-        nextPaginationToken?: string
-    ): Promise<{
-        users: User[];
-        nextPaginationToken?: string | undefined;
-    }> {
-        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getUsersNewestFirst({ limit, nextPaginationToken });
-    }
-
-    /**
-     * @deprecated Use supertokens.getUserCount(...) function instead IF using core version >= 3.5
-     *   */
-    static getUserCount(): Promise<number> {
-        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getUserCount();
-    }
-
     static updateEmailOrPassword(input: { userId: string; email?: string; password?: string }) {
         return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.updateEmailOrPassword(input);
     }
@@ -153,21 +120,6 @@ export let isEmailVerified = Wrapper.isEmailVerified;
 export let revokeEmailVerificationTokens = Wrapper.revokeEmailVerificationTokens;
 
 export let unverifyEmail = Wrapper.unverifyEmail;
-
-/**
- * @deprecated Use supertokens.getUsersOldestFirst(...) function instead IF using core version >= 3.5
- *   */
-export let getUsersOldestFirst = Wrapper.getUsersOldestFirst;
-
-/**
- * @deprecated Use supertokens.getUsersNewestFirst(...) function instead IF using core version >= 3.5
- *   */
-export let getUsersNewestFirst = Wrapper.getUsersNewestFirst;
-
-/**
- * @deprecated Use supertokens.getUserCount(...) function instead IF using core version >= 3.5
- *   */
-export let getUserCount = Wrapper.getUserCount;
 
 export let updateEmailOrPassword = Wrapper.updateEmailOrPassword;