+++ lib/ts/recipe/session/index.ts
@@ -23,7 +23,6 @@ import {
     APIOptions,
 } from "./types";
 import Recipe from "./recipe";
-import framework from "./framework";
 import type { SessionRequest } from "../../framework/express";
 
 // For Express
@@ -69,21 +68,11 @@ export default class SessionWrapper {
         return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.revokeMultipleSessions({ sessionHandles });
     }
 
-    /** @deprecated Use getSessionInformation instead IF using core version >= 3.5 **/
-    static getSessionData(sessionHandle: string) {
-        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getSessionData({ sessionHandle });
-    }
-
     static updateSessionData(sessionHandle: string, newSessionData: any) {
         return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.updateSessionData({
             sessionHandle,
             newSessionData,
         });
-    }
-
-    /** @deprecated Use getSessionInformation instead IF using core version >= 3.5 **/
-    static getJWTPayload(sessionHandle: string) {
-        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getJWTPayload({ sessionHandle });
     }
 
     static updateJWTPayload(sessionHandle: string, newJWTPayload: any) {
@@ -109,20 +98,9 @@ export let revokeSession = SessionWrapper.revokeSession;
 
 export let revokeMultipleSessions = SessionWrapper.revokeMultipleSessions;
 
-/** @deprecated Use getSessionInformation instead IF using core version >= 3.5 **/
-export let getSessionData = SessionWrapper.getSessionData;
-
 export let updateSessionData = SessionWrapper.updateSessionData;
 
-/** @deprecated Use getSessionInformation instead IF using core version >= 3.5 **/
-export let getJWTPayload = SessionWrapper.getJWTPayload;
-
 export let updateJWTPayload = SessionWrapper.updateJWTPayload;
-
-/**
- * @deprecated
- */
-export let verifySession = framework.express.verifySession;
 
 export let Error = SessionWrapper.Error;