+++ lib/ts/recipe/thirdparty/types.ts
@@ -67,18 +67,6 @@ export type TypeInputSetSessionDataForSession = (
     action: "signin" | "signup"
 ) => Promise<{ [key: string]: any } | undefined>;
 
-export type TypeInputSessionFeature = {
-    /**
-     * @deprecated Use override functions instead for >= v6.0
-     *   */
-    setJwtPayload?: TypeInputSetJwtPayloadForSession;
-
-    /**
-     * @deprecated Use override functions instead for >= v6.0
-     *   */
-    setSessionData?: TypeInputSetSessionDataForSession;
-};
-
 const InputSessionFeatureSchema = {
     type: "object",
     properties: {
@@ -86,11 +74,6 @@ const InputSessionFeatureSchema = {
         setSessionData: TypeAny,
     },
     additionalProperties: false,
-};
-
-export type TypeNormalisedInputSessionFeature = {
-    setJwtPayload: TypeInputSetJwtPayloadForSession;
-    setSessionData: TypeInputSetSessionDataForSession;
 };
 
 export type TypeInputEmailVerificationFeature = {
@@ -127,7 +110,6 @@ export type TypeNormalisedInputSignInAndUp = {
 };
 
 export type TypeInput = {
-    sessionFeature?: TypeInputSessionFeature;
     signInAndUpFeature: TypeInputSignInAndUp;
     emailVerificationFeature?: TypeInputEmailVerificationFeature;
     override?: {
@@ -153,7 +135,6 @@ export const InputSchema = {
 };
 
 export type TypeNormalisedInput = {
-    sessionFeature: TypeNormalisedInputSessionFeature;
     signInAndUpFeature: TypeNormalisedInputSignInAndUp;
     emailVerificationFeature: TypeInputEmailVerification;
     override: {
@@ -172,33 +153,6 @@ export interface RecipeInterface {
     getUsersByEmail(input: { email: string }): Promise<User[]>;
 
     getUserByThirdPartyInfo(input: { thirdPartyId: string; thirdPartyUserId: string }): Promise<User | undefined>;
-
-    /**
-     * @deprecated Please do not override this function
-     *   */
-    getUsersOldestFirst(input: {
-        limit?: number;
-        nextPaginationToken?: string;
-    }): Promise<{
-        users: User[];
-        nextPaginationToken?: string;
-    }>;
-
-    /**
-     * @deprecated Please do not override this function
-     *   */
-    getUsersNewestFirst(input: {
-        limit?: number;
-        nextPaginationToken?: string;
-    }): Promise<{
-        users: User[];
-        nextPaginationToken?: string;
-    }>;
-
-    /**
-     * @deprecated Please do not override this function
-     *   */
-    getUserCount(): Promise<number>;
 
     signInUp(input: {
         thirdPartyId: string;