+++ lib/ts/recipe/thirdpartyemailpassword/types.ts
@@ -80,18 +80,6 @@ export type TypeInputSetSessionDataForSession = (
     action: "signin" | "signup"
 ) => Promise<{ [key: string]: any } | undefined>;
 
-export type TypeInputSessionFeature = {
-    /**
-     * @deprecated Use override functions instead for >= v6.0
-     *   */
-    setJwtPayload?: TypeInputSetJwtPayloadForSession;
-
-    /**
-     * @deprecated Use override functions instead for >= v6.0
-     *   */
-    setSessionData?: TypeInputSetSessionDataForSession;
-};
-
 const InputSessionFeatureSchema = {
     type: "object",
     properties: {
@@ -99,11 +87,6 @@ const InputSessionFeatureSchema = {
         setSessionData: TypeAny,
     },
     additionalProperties: false,
-};
-
-export type TypeNormalisedInputSessionFeature = {
-    setJwtPayload: TypeInputSetJwtPayloadForSession;
-    setSessionData: TypeInputSetSessionDataForSession;
 };
 
 export type TypeInputSignUp = {
@@ -149,7 +132,6 @@ const InputEmailVerificationFeatureSchema = {
 };
 
 export type TypeInput = {
-    sessionFeature?: TypeInputSessionFeature;
     signUpFeature?: TypeInputSignUp;
     providers?: TypeProvider[];
     resetPasswordUsingTokenFeature?: TypeInputResetPasswordUsingTokenFeature;
@@ -178,7 +160,6 @@ export const InputSchema = {
 };
 
 export type TypeNormalisedInput = {
-    sessionFeature: TypeNormalisedInputSessionFeature;
     signUpFeature: TypeNormalisedInputSignUp;
     providers: TypeProvider[];
     resetPasswordUsingTokenFeature?: TypeInputResetPasswordUsingTokenFeature;
@@ -199,33 +180,6 @@ export interface RecipeInterface {
     getUsersByEmail(input: { email: string }): Promise<User[]>;
 
     getUserByThirdPartyInfo(input: { thirdPartyId: string; thirdPartyUserId: string }): Promise<User | undefined>;
-
-    /**
-     * @deprecated Please do not override this function
-     *   */
-    getUsersOldestFirst(input: {
-        limit?: number;
-        nextPaginationToken?: string;
-    }): Promise<{
-        users: User[];
-        nextPaginationToken?: string;
-    }>;
-
-    /**
-     * @deprecated Please do not override this function
-     *   */
-    getUsersNewestFirst(input: {
-        limit?: number;
-        nextPaginationToken?: string;
-    }): Promise<{
-        users: User[];
-        nextPaginationToken?: string;
-    }>;
-
-    /**
-     * @deprecated Please do not override this function
-     *   */
-    getUserCount(): Promise<number>;
 
     signInUp(input: {
         thirdPartyId: string;
@@ -251,11 +205,6 @@ export interface RecipeInterface {
         email: string;
         password: string;
     }): Promise<{ status: "OK"; user: User } | { status: "WRONG_CREDENTIALS_ERROR" }>;
-
-    /**
-     * @deprecated Please do not override this function
-     *   */
-    getUserByEmail(input: { email: string }): Promise<User | undefined>;
 
     createResetPasswordToken(input: {
         userId: string;